iot4-nfc-tag-identification/backend/requirements.txt
adafruit-circuitpython-pn532
lgpio
pycardano
python-dotenv
requests
fastapi
uvicorn[standard]
websockets
